correct bugs
