+++ affectSimulator/pom.xml
@@ -24,8 +24,6 @@
             <groupId>org.eclipse.paho</groupId>
             <artifactId>org.eclipse.paho.client.mqttv3</artifactId>
             <version>1.2.5</version>
-            <scope>system</scope>
-            <systemPath>${project.basedir}/../libs/org.eclipse.paho.client.mqttv3-1.2.5.jar</systemPath>
         </dependency>
     </dependencies>
 </project>
