--- affectSimulator/src/main/java/ui/SliderPanel.java
@@ -1,55 +1,0 @@
-package com.company.ui;
-
-import javax.swing.*;
-import javax.swing.border.EmptyBorder;
-import java.awt.*;
-import java.util.Arrays;
-import java.util.List;
-import java.util.stream.Collectors;
-import com.company.SliderListener;
-
-public class SliderPanel extends JPanel {
-    private List<JSlider> sliders;
-    private SliderListener sliderListener;
-
-    public SliderPanel(SliderListener sliderListener) {
-        this.sliderListener = sliderListener;
-        List<String> names = Arrays.asList("Focus", "Stress", "Engagement", "Excitement", "Interest");
-
-        setLayout(new GridLayout(names.size(), 1, 10, 10));
-        setBorder(new EmptyBorder(20, 20, 20, 20));
-
-        sliders = names.stream()
-                .map(this::createSliderWithLabel)  // 使用辅助方法创建带标签的滑块
-                .collect(Collectors.toList());
-    }
-
-    private JSlider createSliderWithLabel(String name) {
-        JSlider slider = new JSlider(0, 100, 50);
-        setupSlider(slider, name);
-
-        JLabel label = new JLabel(name, JLabel.LEFT);
-        JPanel rowPanel = new JPanel(new BorderLayout());
-        rowPanel.add(label, BorderLayout.WEST);
-        rowPanel.add(slider, BorderLayout.CENTER);
-        add(rowPanel);
-
-        return slider;
-    }
-
-    private void setupSlider(JSlider slider, String name) {
-        slider.setMajorTickSpacing(25);
-        slider.setPaintTicks(true);
-        slider.setPaintLabels(true);
-
-        slider.addChangeListener(e -> {
-            if (!slider.getValueIsAdjusting()) { // 仅在滑动停止时更新
-                sliderListener.setSliderValue(name, slider.getValue());
-            }
-        });
-    }
-
-    public List<JSlider> getSliders() {
-        return sliders;
-    }
-}
